Separate Resolver concerns.
Concern 1: The resolver finds the correct value for a given set of parameter properties. This is ConfigResolver

Concern 2: The resolver is responsible for converting the ConfigLoader's values into a new set of values and sending updates.

Driving force for this was that the server side we need to evaluate feature flags. To do this, we need to run the resolver with a map of Config. The ConfigStore interface is now a much simpler wrapper on a datastore.